requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.10.0; python_version >= '3.9'
asgiref==3.9.1; python_version >= '3.9'
certifi==2025.8.3; python_version >= '3.7'
cffi==2.0.0; python_full_version >= '3.9' and platform_python_implementation != 'PyPy'
cryptography==46.0.1
deprecation==2.1.0
django==5.2.6; python_version >= '3.10'
django-cors-headers==4.9.0; python_version >= '3.9'
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
exceptiongroup==1.3.0; python_version < '3.11'
h11==0.16.0; python_version >= '3.8'
h2==4.3.0
hpack==4.1.0; python_version >= '3.9'
httpcore==1.0.9; python_version >= '3.8'
httpx[http2]==0.28.1; python_version >= '3.8'
hyperframe==6.1.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
packaging==25.0; python_version >= '3.8'
pillow==11.3.0; python_version >= '3.9'
postgrest==2.19.0; python_version >= '3.9'
pycparser==2.23; implementation_name != 'PyPy'
pydantic==2.11.9; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyjwt[crypto]==2.10.1; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
realtime==2.19.0; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
storage3==2.19.0; python_version >= '3.9'
strenum==0.4.15; python_full_version < '3.11'
supabase==2.19.0; python_version >= '3.9'
supabase-auth==2.19.0; python_version >= '3.9'
supabase-functions==2.19.0; python_version >= '3.9'
typing-extensions==4.15.0; python_version < '3.11'
typing-inspection==0.4.1; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
